requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.1
fonttools==4.62.0
h11==0.16.0
idna==3.11
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
regex==2026.2.28
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
starlette==0.52.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
